fix(iOS): 修复 ZipStoreWriter 大量 "Ambiguous use of littleEndian(_:into:)" 编译错误
根因：littleEndian(_:into:) 存在 UInt32 与 UInt16 两个重载，
而 local file header / central directory / EOCD 写入处的整数字面量
（如 0x04034b50、0、0x031E、0x06054b50）未显式标注类型，Swift
类型推断无法在两个重载间抉择，每处字面量调用均产生一条歧义
编译错误（约 20 处）。

修复：所有传入字面量的调用统一显式转换为对应宽度：
- 4 字节字段（签名/大小/CRC/偏移）：UInt32(0x04034b50) 等
- 2 字节字段（版本/标志/时间日期/长度/属性）：UInt16(0) 等

传入已定型变量（crc、version、flags、method、localOffset 等）
的调用不受影响。ZIP 二进制布局与字节序无任何变化。

另全库扫描：其余源文件（XlsxWriter 等）无同类的数值重载
歧义模式。

+++ iOS/WatermarkCamera/Services/ZipWriter.swift
@@ -76,17 +76,18 @@ final class ZipStoreWriter {
         let version: UInt16 = 20
 
         var local = Data()
-        Self.littleEndian(0x04034b50, into: &local)          // local file header sig
+        // 注意：字面量必须显式标注 UInt32/UInt16，否则与两个重载产生歧义
+        Self.littleEndian(UInt32(0x04034b50), into: &local) // local file header sig
         Self.littleEndian(version, into: &local)             // version needed
         Self.littleEndian(flags, into: &local)               // flags
         Self.littleEndian(method, into: &local)              // method: 0 = store
-        Self.littleEndian(0, into: &local)                   // mod time
-        Self.littleEndian(0, into: &local)                   // mod date
+        Self.littleEndian(UInt16(0), into: &local)           // mod time
+        Self.littleEndian(UInt16(0), into: &local)           // mod date
         Self.littleEndian(crc, into: &local)                 // crc32
         Self.littleEndian(UInt32(data.count), into: &local)  // compressed size
         Self.littleEndian(UInt32(data.count), into: &local)  // uncompressed size
         Self.littleEndian(UInt16(nameData.count), into: &local) // name len
-        Self.littleEndian(0, into: &local)                   // extra len
+        Self.littleEndian(UInt16(0), into: &local)           // extra len
         local.append(nameData)
 
         try handle.write(contentsOf: local)
@@ -94,22 +95,22 @@ final class ZipStoreWriter {
 
         // 中央目录条目
         var central = Data()
-        Self.littleEndian(0x02014b50, into: &central)        // central file header sig
-        Self.littleEndian(0x031E, into: &central)            // version made by
+        Self.littleEndian(UInt32(0x02014b50), into: &central) // central file header sig
+        Self.littleEndian(UInt16(0x031E), into: &central)   // version made by
         Self.littleEndian(version, into: &central)           // version needed
         Self.littleEndian(flags, into: &central)             // flags
         Self.littleEndian(method, into: &central)            // method
-        Self.littleEndian(0, into: &central)                 // mod time
-        Self.littleEndian(0, into: &central)                 // mod date
+        Self.littleEndian(UInt16(0), into: &central)        // mod time
+        Self.littleEndian(UInt16(0), into: &central)        // mod date
         Self.littleEndian(crc, into: &central)               // crc32
         Self.littleEndian(UInt32(data.count), into: &central) // compressed size
         Self.littleEndian(UInt32(data.count), into: &central) // uncompressed size
         Self.littleEndian(UInt16(nameData.count), into: &central) // name len
-        Self.littleEndian(0, into: &central)                 // extra len
-        Self.littleEndian(0, into: &central)                 // comment len
-        Self.littleEndian(0, into: &central)                 // disk number
-        Self.littleEndian(0, into: &central)                 // internal attrs
-        Self.littleEndian(0, into: &central)                 // external attrs
+        Self.littleEndian(UInt16(0), into: &central)        // extra len
+        Self.littleEndian(UInt16(0), into: &central)        // comment len
+        Self.littleEndian(UInt16(0), into: &central)        // disk number
+        Self.littleEndian(UInt16(0), into: &central)        // internal attrs
+        Self.littleEndian(UInt16(0), into: &central)        // external attrs
         Self.littleEndian(localOffset, into: &central)       // local header offset
         central.append(nameData)
 
@@ -127,7 +128,7 @@ final class ZipStoreWriter {
         let cdSize = UInt32(centralDirectory.count)
 
         var eocd = Data()
-        Self.littleEndian(0x06054b50, into: &eocd)    // EOCD sig
+        Self.littleEndian(UInt32(0x06054b50), into: &eocd) // EOCD sig
         Self.littleEndian(UInt16(0), into: &eocd)     // disk number
         Self.littleEndian(UInt16(0), into: &eocd)     // cd start disk
         Self.littleEndian(UInt16(min(entryCount, 0xFFFF)), into: &eocd) // entries on this disk
